Symfony 3.3 compatibility refacto.
I got problems to use Form/Type/RelatedToManyMetadataType cause of
dependency injection problems. The MetadataRepository system was out of date.

Updated services.yml to use new symfony autowire and autoconfig.
Removed useless MetadataManager which was just a repository wrapper.
Used Sensio Lab php cs fixer.
