+++ DependencyInjection/Compiler/MetadataNamespaceCompilerPass.php
@@ -7,8 +7,8 @@
 
 namespace IDCI\Bundle\SimpleMetadataBundle\DependencyInjection\Compiler;
 
-use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\DefinitionDecorator;
 
 class MetadataNamespaceCompilerPass implements CompilerPassInterface
