+++ Form/Type/JsonType.php
@@ -7,10 +7,10 @@
 
 namespace IDCI\Bundle\SimpleMetadataBundle\Form\Type;
 
+use IDCI\Bundle\SimpleMetadataBundle\Form\DataTransformer\JsonToArrayTransformer;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type as Types;
 use Symfony\Component\Form\FormBuilderInterface;
-use IDCI\Bundle\SimpleMetadataBundle\Form\DataTransformer\JsonToArrayTransformer;
 
 class JsonType extends AbstractType
 {
